siva: move repositoryIterator to its own file and other minor changes
